EPMLABSBRN-306 rename class CsvMappingIteratorParser -> CsvParser <eom>
